+++ src/gdb/exec_result/recv/asm_insns.rs
@@ -12,6 +12,9 @@ pub fn recv_exec_result_asm_insns(state: &mut State, asm: &String) {
     if let Written::AsmAtPc = last_written {
         state.asm = parse_asm_insns_values(asm).clone();
     }
+    if let Written::SymbolDisassembly(_name) = &last_written {
+        state.symbol_asm = parse_asm_insns_values(asm).clone();
+    }
     if let Written::SymbolAtAddrRegister((base_reg, _n)) = &last_written {
         for RegisterStorage { name: _, register, deref } in state.registers.iter_mut() {
             if let Some(reg) = register {