+++ src/gdb/exec_result/recv/mod.rs
@@ -3,4 +3,5 @@ pub mod changed_registers;
 pub mod register_names;
 pub mod register_values;
 pub mod result_memory;
+pub mod symbols;
 pub mod value;